Add postPage router
