+++ client/src/PostingPage.js
@@ -1,0 +1,9 @@
+function Posting(){
+    return(
+        <div>
+            test
+        </div>
+    )
+}
+
+export default Posting;
