+++ client/src/style.scss
@@ -3,7 +3,7 @@ body{
     padding-bottom: 70px;
 }
 
-.App{
+.main{
     position: relative;
     margin:auto;
     overflow: hidden;
